requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==1.7.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
